{{instigator_py.project_slug}}/requirements/base.txt
pytz==2023.3  # https://github.com/stub42/pytz
python-slugify==8.0.1  # https://github.com/un33k/python-slugify
Pillow==9.5.0  # https://github.com/python-pillow/Pillow
{%- if instigator_py.frontend_pipeline == 'Django Compressor' %}
{%- if instigator_py.windows == 'y' and instigator_py.use_docker == 'n' %}
rcssmin==1.1.0 --install-option="--without-c-extensions"  # https://github.com/ndparker/rcssmin
{%- else %}
rcssmin==1.1.1  # https://github.com/ndparker/rcssmin
{%- endif %}
{%- endif %}
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
{%- if instigator_py.use_whitenoise == 'y' %}
whitenoise==6.4.0  # https://github.com/evansd/whitenoise
{%- endif %}
redis==4.5.4  # https://github.com/redis/redis-py
{%- if instigator_py.use_docker == "y" or instigator_py.windows == "n" %}
hiredis==2.2.2  # https://github.com/redis/hiredis-py
{%- endif %}
{%- if instigator_py.use_celery == "y" %}
celery==5.2.7  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.5.0  # https://github.com/celery/django-celery-beat
{%- if instigator_py.use_docker == 'y' %}
flower==1.2.0  # https://github.com/mher/flower
{%- endif %}
{%- endif %}
{%- if instigator_py.use_async == 'y' %}
uvicorn[standard]==0.21.1  # https://github.com/encode/uvicorn
{%- endif %}

# Django
# ------------------------------------------------------------------------------
django==3.2.15  # pyup: < 3.2.15  # https://www.djangoproject.com/
django-environ==0.10.0  # https://github.com/joke2k/django-environ
django-model-utils==4.3.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.54.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==2.0  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.7  # https://github.com/django-crispy-forms/crispy-bootstrap5
{%- if instigator_py.frontend_pipeline == 'Django Compressor' %}
django-compressor==4.3.1  # https://github.com/django-compressor/django-compressor
{%- endif %}
django-redis==5.2.0  # https://github.com/jazzband/django-redis
{%- if instigator_py.use_drf == 'y' %}
# Django REST Framework
djangorestframework==3.14.0  # https://github.com/encode/django-rest-framework
# DRF-spectacular for api documentation
drf-spectacular==0.26.2  # https://github.com/tfranzel/drf-spectacular
{%- endif %}
{%- if instigator_py.frontend_pipeline == 'Webpack' %}
django-webpack-loader==1.8.1  # https://github.com/django-webpack/django-webpack-loader
{%- endif %}
{%- if instigator_py.use_graphene == 'y' %}
# Graphene
graphene-django==3.0.0  # https://github.com/graphql-python/graphene-django
django-graphql-jwt==0.3.4  # https://github.com/flavors/django-graphql-jwt
graphene-file-upload==1.3.0  
{%- endif %}
django-filter==22.1  # https://github.com/carltongibson/django-filter
django-cors-headers==3.14.0  # https://github.com/adamchainz/django-cors-headers
{%- if instigator_py.use_drf == "y" %}
dj-rest-auth==3.0.0
djangorestframework-simplejwt==5.2.2
{%- endif %}
{%- if instigator_py.use_tenants == "y" %}
django-tenants==3.4.8
{%- endif %}

